env {LEA}	R Documentation

Environmental input file format for lfmm

Description

Description of the env format. The env format can be used as an input format for the environmental variables in the lfmm function.

Details

The env format has one row for each individual. Each row contains one value for each environmental variable (separated by spaces or tabulations).

Here is an example of an environmental file using the env format with 3 individuals and 2 variables:

0.252477 0.95250639
0.216618 0.10902647
-0.47509 0.07626694

Author(s)

Eric Frichot

See Also

lfmm lfmm2 read.env write.env

[Package LEA version 3.19.8 Index]