read.geno {LEA}	R Documentation

read a file in the geno format

Description

Read a file in the geno format.

Usage

read.geno(input.file)

Arguments

input.file
A character string containing a path to the input file, a genotypic matrix in the geno format.

Value

R
A matrix containing the genotypes with one line for each individual and one column for each SNP.

Author(s)

Eric Frichot

See Also

write.geno geno snmf geno2lfmm lfmm2geno ancestrymap2geno ped2geno vcf2geno

Examples

# tutorial contains a matrix of genotypes R with 1000 SNPs for 165 individuals.
# and a matrix with an environmental variable C.
data("tutorial")

# Write R in a file called "genotypes.geno".
# Create file:    "genotypes.geno".
write.geno(tutorial.R,"genotypes.geno")

# Read the file "genotypes.geno".
R = read.geno("genotypes.geno")

[Package LEA version 3.19.8 Index]