zscore.format {LEA}	R Documentation

Output file format for lfmm

Description

Description of the zscore output format of lfmm.

Details

The zscore format has one row for each SNP. Each row contains three values: The first value is the zscore, the second value is the -log10(pvalue), the third value is the p-value (separated by spaces or tabulations).

Author(s)

Eric Frichot

See Also

lfmm lfmm.data env

[Package LEA version 3.19.8 Index]